feature: Support edge case of sub-sampling in λ
The case is a scenario to compute λ when using the same extensions for both
the directions and λ. If the samples used for λ are not fully contained in
the samples for the directions, additional scalar products than those contained
in the Gram matrix need to be computed. The current solution recomputes all
scalar products, which makes the tests pass.

A future improvement could be to partly recycle scalar products from the Gram
matrix and only compute those that really have not yet been computed.